// src/findjobnu-api/models/CategoryJobCountResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FindjobnuService
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CategoryJobCountResponse
 */
export interface CategoryJobCountResponse {
    /**
     * 
     * @type {number}
     * @memberof CategoryJobCountResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof CategoryJobCountResponse
     */
    name?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CategoryJobCountResponse
     */
    numberOfJobs?: number;
}

/**
 * Check if a given object implements the CategoryJobCountResponse interface.
 */
export function instanceOfCategoryJobCountResponse(value: object): value is CategoryJobCountResponse {
    return true;
}

export function CategoryJobCountResponseFromJSON(json: any): CategoryJobCountResponse {
    return CategoryJobCountResponseFromJSONTyped(json, false);
}

export function CategoryJobCountResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CategoryJobCountResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'numberOfJobs': json['numberOfJobs'] == null ? undefined : json['numberOfJobs'],
    };
}

export function CategoryJobCountResponseToJSON(json: any): CategoryJobCountResponse {
    return CategoryJobCountResponseToJSONTyped(json, false);
}

export function CategoryJobCountResponseToJSONTyped(value?: CategoryJobCountResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'numberOfJobs': value['numberOfJobs'],
    };
}