// src/findjobnu-api/models/JobStatisticsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FindjobnuService
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CategoryJobCountResponse } from './CategoryJobCountResponse';
import {
    CategoryJobCountResponseFromJSON,
    CategoryJobCountResponseFromJSONTyped,
    CategoryJobCountResponseToJSON,
    CategoryJobCountResponseToJSONTyped,
} from './CategoryJobCountResponse';

/**
 * 
 * @export
 * @interface JobStatisticsResponse
 */
export interface JobStatisticsResponse {
    /**
     * 
     * @type {Array<CategoryJobCountResponse>}
     * @memberof JobStatisticsResponse
     */
    topCategories?: Array<CategoryJobCountResponse> | null;
    /**
     * 
     * @type {Array<CategoryJobCountResponse>}
     * @memberof JobStatisticsResponse
     */
    topCategoriesLastWeek?: Array<CategoryJobCountResponse> | null;
    /**
     * 
     * @type {number}
     * @memberof JobStatisticsResponse
     */
    totalJobs?: number;
    /**
     * 
     * @type {number}
     * @memberof JobStatisticsResponse
     */
    newJobsLastWeek?: number;
    /**
     * 
     * @type {number}
     * @memberof JobStatisticsResponse
     */
    newJobsLastMonth?: number;
}

/**
 * Check if a given object implements the JobStatisticsResponse interface.
 */
export function instanceOfJobStatisticsResponse(value: object): value is JobStatisticsResponse {
    return true;
}

export function JobStatisticsResponseFromJSON(json: any): JobStatisticsResponse {
    return JobStatisticsResponseFromJSONTyped(json, false);
}

export function JobStatisticsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): JobStatisticsResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'topCategories': json['topCategories'] == null ? undefined : ((json['topCategories'] as Array<any>).map(CategoryJobCountResponseFromJSON)),
        'topCategoriesLastWeek': json['topCategoriesLastWeek'] == null ? undefined : ((json['topCategoriesLastWeek'] as Array<any>).map(CategoryJobCountResponseFromJSON)),
        'totalJobs': json['totalJobs'] == null ? undefined : json['totalJobs'],
        'newJobsLastWeek': json['newJobsLastWeek'] == null ? undefined : json['newJobsLastWeek'],
        'newJobsLastMonth': json['newJobsLastMonth'] == null ? undefined : json['newJobsLastMonth'],
    };
}

export function JobStatisticsResponseToJSON(json: any): JobStatisticsResponse {
    return JobStatisticsResponseToJSONTyped(json, false);
}

export function JobStatisticsResponseToJSONTyped(value?: JobStatisticsResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'topCategories': value['topCategories'] == null ? undefined : ((value['topCategories'] as Array<any>).map(CategoryJobCountResponseToJSON)),
        'topCategoriesLastWeek': value['topCategoriesLastWeek'] == null ? undefined : ((value['topCategoriesLastWeek'] as Array<any>).map(CategoryJobCountResponseToJSON)),
        'totalJobs': value['totalJobs'],
        'newJobsLastWeek': value['newJobsLastWeek'],
        'newJobsLastMonth': value['newJobsLastMonth'],
    };
}